Move get_event_impact and make_inspection_priority
