+++ sc/app/functions.py
@@ -23,6 +23,7 @@ from inventory import (
 )
 from orm import *
 import pdf
+from impact import get_event_impact, make_inspection_priority
 from jsonencoders import AlchemyEncoder, makeImpactGeoJSONDict, saveImpactGeoJson
 from notifications import NotificationBuilder, TemplateManager, Mailer
 from productdownload import geo_json, download_scenario
@@ -304,46 +305,6 @@ def process_shakemaps(shakemaps=None, session=None, scenario=False):
         
         session.commit()
 
-def make_inspection_priority(facility=None,
-                          shakemap=None,
-                          grid=None):
-    '''
-    Determines inspection priorities for the input facility
-    
-    Args:
-        facility (Facility): A facility to be processed
-        shakemap (ShakeMap): The ShakeMap which is associated with the shaking
-        grid (ShakeMapGrid): The grid built from the ShakeMap
-        notifications (list): List of Notification objects which should be associated with the shaking
-        
-    Returns:
-        dict: A dictionary with all the parameters needed to make a FacilityShaking entry in the database
-        ::
-            fac_shaking = {'gray': PDF Value,
-                           'green': PDF Value,
-                           'yellow': PDF Value,
-                           'orange': PDF Value,
-                           'red': PDF Value,
-                           'metric': which metric is used to compute PDF values,
-                           'facility_id': shakecast_id of the facility that's shaking,
-                           'shakemap_id': shakecast_id of the associated ShakeMap,
-                           '_shakecast_id': ID for the FacilityShaking entry that will be created,
-                           'update': bool -- True if an ID already exists for this FacilityShaking,
-                           'alert_level': string ('gray', 'green', 'yellow' ...),
-                           'weight': float that determines inspection priority,
-                           'notifications': list of notifications associated with this shaking}
-    '''
-    
-    # get the largest shaking level affecting the facility
-    shaking_point = grid.max_shaking(facility=facility)
-    if shaking_point is None:
-        return False
-    
-    # use the max shaking value to create fragility curves for the
-    # damage states
-    fac_shaking = facility.make_alert_level(shaking_point=shaking_point,
-                                            shakemap=shakemap)
-    return fac_shaking
     
 def new_event_notification(notifications=None,
                            scenario=False):
@@ -676,17 +637,3 @@ def run_scenario(shakemap_id=None, session=None):
             'error': error,
             'log': 'Run scenario: ' + shakemap_id}
 
-def get_event_impact(shakemap):
-    impact_sum = {'gray': 0,
-             'green': 0,
-             'yellow': 0,
-             'orange': 0,
-             'red': 0}
-
-    fac_shaking = shakemap.facility_shaking
-    
-    for s in fac_shaking:
-        # record number of facs at each alert level
-        impact_sum[s.alert_level] += 1
-
-    return impact_sum
